backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.27.0

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.2.2
alembic>=1.13.1

# Authentication
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Google APIs
google-api-python-client>=2.115.0

# AI and Document Processing
anthropic>=0.40.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
pandas>=2.2.0
openpyxl>=3.1.2
pytesseract>=0.3.10

# Storage
boto3>=1.34.34

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0
